issue: 3788369 Move public types definitions to xlio_types.h
xlio.h and xlio_extra.h provides public API for different scenarios:
explicit linkage and LD_PRELOAD respectively. Therefore, usually, it
makes sense to use a single header. However, the headers have common
types definitions. Move them to a separate xlio_types.h instead of
including xlio_extra.h from xlio.h.
