+++ src/core/xlio.h
@@ -40,8 +40,9 @@
 #include <stdint.h>
 #include <sys/epoll.h>
 #include <sys/stat.h>
+#include <sys/time.h>
 
-#include "xlio_extra.h"
+#include "xlio_types.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -387,14 +388,9 @@ int xlio_socketxtreme_free_buff(struct xlio_buff_t *buff);
 
 /*
  * XLIO Socket API
+ *
+ * This is performance-oriented event based API.
  */
-
-typedef uintptr_t xlio_poll_group_t;
-typedef uintptr_t xlio_socket_t;
-
-struct xlio_buf {
-    uint64_t userdata;
-};
 
 /*
  * XLIO initialization.
@@ -406,79 +402,7 @@ struct xlio_buf {
  * If set, memory_cb() notifies about memory blocks which zerocopy RX buffers can point to.
  * Current implementation allocates a single memory block and does it within xlio_init_ex() context.
  */
-
-/*
- * Memory callback.
- *
- * XLIO calls the callback each time XLIO allocates a memory region which can be used for RX
- * buffers. User can use this information to prepare the memory for some logic in the future.
- * Zerocopy RX interface provides pointers to such memory.
- *
- * Current XLIO implementation does a single allocation for buffers.
- */
-typedef void (*xlio_memory_cb_t)(void *addr, size_t len, size_t hugepage_size);
-
-struct xlio_init_attr {
-    unsigned flags;
-    xlio_memory_cb_t memory_cb;
-};
-
 int xlio_init_ex(const struct xlio_init_attr *attr);
-
-/*
- * Socket callbacks.
- */
-
-enum {
-    /* TCP connection established. */
-    XLIO_SOCKET_EVENT_ESTABLISHED = 1,
-    /* Socket terminated and no further events are possible. */
-    XLIO_SOCKET_EVENT_TERMINATED,
-    /* Passive close. */
-    XLIO_SOCKET_EVENT_CLOSED,
-    /* An error occurred, see the error code value. */
-    XLIO_SOCKET_EVENT_ERROR,
-};
-
-/*
- * Socket event callback.
- *
- * May be called from xlio_poll_group_poll() context.
- * In the callback context, send operation is allowed only for the ESTABLISHED event.
- * Argument value holds the error code for the ERROR event and 0 for other events.
- *
- * List of possible error code values:
- * ECONNABORTED - connection aborted by local side
- * ECONNRESET - connection reset by remote side
- * ECONNREFUSED - connection refused by remote side during TCP handshake
- * ETIMEDOUT - connection timed out due to keepalive, user timeout option or TCP handshake timeout
- */
-typedef void (*xlio_socket_event_cb_t)(xlio_socket_t, uintptr_t userdata_sq, int event, int value);
-
-/*
- * Zerocopy completion event.
- *
- * May be called from the following contexts:
- *  - xlio_poll_group_poll() - likely
- *  - xlio_socket_send() - can happen only if data is flushed
- *  - xlio_socket_flush() / xlio_poll_group_flush()
- *  - xlio_socket_destroy()
- *
- * In the callback context, send operation is allowed unless the socket is under destruction.
- */
-typedef void (*xlio_socket_comp_cb_t)(xlio_socket_t, uintptr_t userdata_sq, uintptr_t userdata_op);
-
-/*
- * RX callback.
- *
- * Returns TCP payload upon arrival. Each call returns a single contiguous buffer. The buffer points
- * to memory within a block which is provided by the memory_cb() notification.
- *
- * xlio_buf is a descriptor of the buffer which must be returned to XLIO. During user ownership,
- * they may use the uninitialized field in the structure.
- */
-typedef void (*xlio_socket_rx_cb_t)(xlio_socket_t, uintptr_t userdata_sq, void *data, size_t len,
-                                    struct xlio_buf *buf);
 
 /*
  * XLIO polling groups.
@@ -494,19 +418,6 @@ typedef void (*xlio_socket_rx_cb_t)(xlio_socket_t, uintptr_t userdata_sq, void *
  *  - Reduce the number of different network interfaces within a group to minimum. This will
  *    optimize the HW objects utilization. However, maintaining extra groups can have an overhead.
  */
-
-/* Sockets and rings will be protected with locks regardless of XLIO configuration. */
-#define XLIO_GROUP_FLAG_SAFE 0x1
-/* Group will keep dirty sockets to be flushed with xlio_poll_group_flush(). */
-#define XLIO_GROUP_FLAG_DIRTY 0x2
-
-struct xlio_poll_group_attr {
-    unsigned flags;
-
-    xlio_socket_event_cb_t socket_event_cb;
-    xlio_socket_comp_cb_t socket_comp_cb;
-    xlio_socket_rx_cb_t socket_rx_cb;
-};
 
 int xlio_poll_group_create(const struct xlio_poll_group_attr *attr, xlio_poll_group_t *group_out);
 int xlio_poll_group_destroy(xlio_poll_group_t group);
@@ -528,13 +439,6 @@ void xlio_poll_group_poll(xlio_poll_group_t group);
  *  - Bonding is not supported
  */
 
-struct xlio_socket_attr {
-    unsigned flags;
-    int domain; /* AF_INET or AF_INET6 */
-    xlio_poll_group_t group;
-    uintptr_t userdata_sq;
-};
-
 /* Forward declaration. */
 struct ibv_pd;
 
@@ -545,8 +449,6 @@ int xlio_socket_setsockopt(xlio_socket_t sock, int level, int optname, const voi
 int xlio_socket_bind(xlio_socket_t sock, const struct sockaddr *addr, socklen_t addrlen);
 int xlio_socket_connect(xlio_socket_t sock, const struct sockaddr *to, socklen_t tolen);
 struct ibv_pd *xlio_socket_get_pd(xlio_socket_t sock);
-
-int xlio_socket_fd(xlio_socket_t sock);
 
 /*
  * TX flow.
@@ -569,17 +471,6 @@ int xlio_socket_fd(xlio_socket_t sock);
  *    it's better to avoid using them both.
  */
 
-/* Flush socket after queueing the data. */
-#define XLIO_SOCKET_SEND_FLAG_FLUSH 0x1
-/* Copy user data to the internal buffers instead of taking ownership. */
-#define XLIO_SOCKET_SEND_FLAG_INLINE 0x2
-
-struct xlio_socket_send_attr {
-    unsigned flags;
-    uint32_t mkey;
-    uintptr_t userdata_op;
-};
-
 /* Returns either 0 or -1. The errors, except of ENOMEM, are not recoverable. */
 int xlio_socket_send(xlio_socket_t sock, const void *data, size_t len,
                      const struct xlio_socket_send_attr *attr);
@@ -594,6 +485,12 @@ void xlio_socket_flush(xlio_socket_t sock);
 
 void xlio_socket_buf_free(xlio_socket_t sock, struct xlio_buf *buf);
 void xlio_poll_group_buf_free(xlio_poll_group_t group, struct xlio_buf *buf);
+
+/*
+ * Experimental level API.
+ */
+
+int xlio_socket_fd(xlio_socket_t sock);
 
 #ifdef __cplusplus
 }
